Fix clang compilation
- Add macro for printing size_t independent of OS, processor bits (32/64)
- Deal with comparison of different types
- Add pose traits to their corresponding header files
