+++ libs/base/include/mrpt/utils/CProbabilityDensityFunction.h
@@ -12,19 +12,31 @@
 #include <mrpt/math/CMatrixTemplateNumeric.h>
 #include <mrpt/math/CMatrixFixedNumeric.h>
 #include <mrpt/math/math_frwds.h>
-#include <mrpt/poses/pose_traits.h>
 
 namespace mrpt
 {
+	namespace traits {
+		/**\brief Type Traits for manipulating covariance and information form
+ 		 * Probability Density Function (PDF) classes
+ 		 *
+ 		 * \ingroup poses_grp
+ 		 */
+		template<class T>
+			struct is_inf_type {
+				// false by default
+				static const bool value = false;
+			};
+	}
+
 	namespace utils
 	{
 		/** A generic template for probability density distributions (PDFs).
-		  * This template is used as base for many classes in mrpt::poses
-		  *  Any derived class must implement \a getMean() and a getCovarianceAndMean().
-		  *  Other methods such as \a getMean() or \a getCovariance() are implemented here for convenience.
-		  * \sa mprt::poses::CPosePDF, mprt::poses::CPose3DPDF, mprt::poses::CPointPDF
+		 * This template is used as base for many classes in mrpt::poses
+		 *  Any derived class must implement \a getMean() and a getCovarianceAndMean().
+		 *  Other methods such as \a getMean() or \a getCovariance() are implemented here for convenience.
+		 * \sa mprt::poses::CPosePDF, mprt::poses::CPose3DPDF, mprt::poses::CPointPDF
 		 * \ingroup mrpt_base_grp
- 		  */
+ 		 */
 		template <class TDATA, size_t STATE_LEN>
 		class CProbabilityDensityFunction
 		{
@@ -96,7 +108,7 @@ namespace mrpt
 			 * \note By default this is going to be covariance form. *Inf classes
 			 * (e.g. CPosePDFGaussianInf) store it in information form.
 			 *
-			 * \sa mrpt::poses_traits::is_inf_type
+			 * \sa mrpt::traits::is_inf_type
 			 */
 			virtual bool isInfType() const { return false; }
